CH: Converting avatar to a seperate component
